Code reorg after review
